+++ src/Math/vector/Vector3F.java
@@ -7,7 +7,7 @@ public class Vector3F extends AbstractVector implements Vector {
     public Vector3F() {
     }
 
-    public Vector3F(double[] values) {
+    public Vector3F(float[] values) {
         if (checkLengthInputValues(values)) {
             super.values = values;
             super.size = values.length;
@@ -15,7 +15,7 @@ public class Vector3F extends AbstractVector implements Vector {
     }
 
     public Vector3F(int v1, int v2, int v3) {
-        super.values = new double[3];
+        super.values = new float[3];
 
         super.size = 3;
 
@@ -26,7 +26,7 @@ public class Vector3F extends AbstractVector implements Vector {
 
     public Vector vectorProduct(Vector v1, Vector v2){
         Vector vRes = new Vector3F();
-        double[] tmp = new double[3];
+        float[] tmp = new float[3];
         if(v1.getSize() == v2.getSize() && checkLengthInputValues(v1.getValues())
                 && checkLengthInputValues(v2.getValues())){
             tmp[0] = v1.getValues()[1] * v2.getValues()[2] - v1.getValues()[2] * v2.getValues()[1];
@@ -38,7 +38,7 @@ public class Vector3F extends AbstractVector implements Vector {
     }
 
     @Override
-    protected boolean checkLengthInputValues(double[] values) {
+    protected boolean checkLengthInputValues(float[] values) {
         return values.length == 3;
     }
